competition/requirements.txt
torch
torchvision
tqdm
PyYAML
scikit-learn
numpy
Pillow
open_clip_torch
optuna
